+++ pjrt/common.c
@@ -26,3 +26,24 @@ extern PJRT_NamedValueUnion Extract_PJRT_NamedValue_Union(PJRT_NamedValue *named
     return result;
 }
 
+// Set the corresponding field in the PJRT_NamedValue structure.
+// The one to use is based on named_value->type.
+void Set_PJRT_NamedValue_Union(PJRT_NamedValue *named_value, PJRT_NamedValueUnion split_value) {
+    switch (named_value->type) {
+    case PJRT_NamedValue_kString:
+        named_value->string_value = split_value.string_value;
+        break;
+    case PJRT_NamedValue_kInt64:
+        named_value->int64_value = split_value.int64_value;
+        break;
+    case PJRT_NamedValue_kInt64List:
+        named_value->int64_array_value = split_value.int64_array_value;
+        break;
+    case PJRT_NamedValue_kFloat:
+        named_value->float_value = split_value.float_value;
+        break;
+    case PJRT_NamedValue_kBool:
+        named_value->bool_value = split_value.bool_value;
+        break;
+    }
+}